giving VehicleFormat its own file; fixing imports
