+++ src/main/scala/net/psforever/objects/definition/converter/UtilityVehicleConverter.scala
@@ -2,7 +2,8 @@
 package net.psforever.objects.definition.converter
 
 import net.psforever.objects.Vehicle
-import net.psforever.packet.game.objectcreate.{UtilityVehicleData, VehicleFormat}
+import net.psforever.packet.game.objectcreate.UtilityVehicleData
+import net.psforever.types.VehicleFormat
 
 class UtilityVehicleConverter extends VehicleConverter {
   override protected def SpecificFormatModifier: VehicleFormat = VehicleFormat.Utility
