+++ src/main/scala/net/psforever/objects/definition/converter/VariantVehicleConverter.scala
@@ -2,7 +2,8 @@
 package net.psforever.objects.definition.converter
 
 import net.psforever.objects.Vehicle
-import net.psforever.packet.game.objectcreate.{VariantVehicleData, VehicleFormat}
+import net.psforever.packet.game.objectcreate.VariantVehicleData
+import net.psforever.types.VehicleFormat
 
 class VariantVehicleConverter extends VehicleConverter {
   override protected def SpecificFormatModifier: VehicleFormat = VehicleFormat.Variant
