+++ src/main/scala/net/psforever/objects/definition/converter/VehicleConverter.scala
@@ -4,7 +4,7 @@ package net.psforever.objects.definition.converter
 import net.psforever.objects.equipment.{Equipment, EquipmentSlot}
 import net.psforever.objects.{PlanetSideGameObject, Vehicle}
 import net.psforever.packet.game.objectcreate._
-import net.psforever.types.{DriveState, PlanetSideGUID}
+import net.psforever.types.{DriveState, PlanetSideGUID, VehicleFormat}
 
 import scala.util.{Failure, Success, Try}
