+++ src/main/scala/net/psforever/packet/game/objectcreate/ObjectClass.scala
@@ -1,6 +1,7 @@
 // Copyright (c) 2017 PSForever
 package net.psforever.packet.game.objectcreate
 
+import net.psforever.types.VehicleFormat
 import scodec.{Attempt, Codec, Err}
 import scodec.codecs._
